+++ CHANGELOG.md
@@ -20,6 +20,12 @@
 * [MIC-6584] Switching terminal now available in ASDKSample app
 * [MIC-6527] NetworkClient unit tests
 * [MIC-6627] Added Tests for PaymentController
+* [MIC-6526] AcquiringAPIClient unit tests
+* [MIC-6801] AcquiringRequestAdapter unit tests
+
+### Fixed
+
+* [MIC-6675] Submit3DSAuthorizationV2 request for 3DS App Based Flow
 
 ## [2.11.2] - 2022-10-13Z